requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html

ipython
keras<3.0.0
tensorflow[and-cuda]==2.14
tqdm
tabulate
natsort
torch==2.2.2+cpu
torchvision==0.17.2+cpu
gitpython
